DBZ-5623 DDL Parsing Error
