+++ debezium-connector-mysql/src/main/java/io/debezium/connector/mysql/antlr/listener/AlterTableParserListener.java
@@ -18,10 +18,8 @@ import org.slf4j.LoggerFactory;
 
 import io.debezium.connector.mysql.antlr.MySqlAntlrDdlParser;
 import io.debezium.ddl.parser.mysql.generated.MySqlParser;
-import io.debezium.ddl.parser.mysql.generated.MySqlParserBaseListener;
 import io.debezium.relational.Column;
 import io.debezium.relational.ColumnEditor;
-import io.debezium.relational.TableEditor;
 import io.debezium.relational.TableId;
 import io.debezium.text.ParsingException;
 
@@ -30,26 +28,20 @@ import io.debezium.text.ParsingException;
  *
  * @author Roman Kuchár <kucharrom@gmail.com>.
  */
-public class AlterTableParserListener extends MySqlParserBaseListener {
+public class AlterTableParserListener extends TableCommonParserListener {
 
     private static final int STARTING_INDEX = 1;
 
     private final static Logger LOG = LoggerFactory.getLogger(AlterTableParserListener.class);
 
-    private final MySqlAntlrDdlParser parser;
-    private final List<ParseTreeListener> listeners;
-
-    private TableEditor tableEditor;
     private ColumnEditor defaultValueColumnEditor;
-    private ColumnDefinitionParserListener columnDefinitionListener;
     private DefaultValueParserListener defaultValueListener;
 
     private List<ColumnEditor> columnEditors;
     private int parsingColumnIndex = STARTING_INDEX;
 
     public AlterTableParserListener(MySqlAntlrDdlParser parser, List<ParseTreeListener> listeners) {
-        this.parser = parser;
-        this.listeners = listeners;
+        super(parser, listeners);
     }
 
     @Override
