+++ debezium-connector-mysql/src/main/java/io/debezium/connector/mysql/antlr/listener/CreateTableParserListener.java
@@ -13,11 +13,8 @@ import org.antlr.v4.runtime.tree.ParseTreeListener;
 
 import io.debezium.connector.mysql.antlr.MySqlAntlrDdlParser;
 import io.debezium.ddl.parser.mysql.generated.MySqlParser;
-import io.debezium.ddl.parser.mysql.generated.MySqlParserBaseListener;
-import io.debezium.relational.Column;
 import io.debezium.relational.ColumnEditor;
 import io.debezium.relational.Table;
-import io.debezium.relational.TableEditor;
 import io.debezium.relational.TableId;
 
 /**
@@ -25,17 +22,10 @@ import io.debezium.relational.TableId;
  *
  * @author Roman Kuchár <kucharrom@gmail.com>.
  */
-public class CreateTableParserListener extends MySqlParserBaseListener {
-
-    private final List<ParseTreeListener> listeners;
-
-    private final MySqlAntlrDdlParser parser;
-    private TableEditor tableEditor;
-    private ColumnDefinitionParserListener columnDefinitionListener;
+public class CreateTableParserListener extends TableCommonParserListener {
 
     public CreateTableParserListener(MySqlAntlrDdlParser parser, List<ParseTreeListener> listeners) {
-        this.parser = parser;
-        this.listeners = listeners;
+        super(parser, listeners);
     }
 
     @Override
@@ -85,48 +75,6 @@ public class CreateTableParserListener extends MySqlParserBaseListener {
             parser.signalCreateTable(tableId, ctx);
         }
         super.exitCopyCreateTable(ctx);
-    }
-
-    @Override
-    public void enterColumnDeclaration(MySqlParser.ColumnDeclarationContext ctx) {
-        parser.runIfNotNull(() -> {
-            String columnName = parser.parseName(ctx.uid());
-            ColumnEditor columnEditor = Column.editor().name(columnName);
-            if (columnDefinitionListener == null) {
-                columnDefinitionListener = new ColumnDefinitionParserListener(tableEditor, columnEditor, parser, listeners);
-                listeners.add(columnDefinitionListener);
-            }
-            else {
-                columnDefinitionListener.setColumnEditor(columnEditor);
-            }
-        }, tableEditor);
-        super.enterColumnDeclaration(ctx);
-    }
-
-    @Override
-    public void exitColumnDeclaration(MySqlParser.ColumnDeclarationContext ctx) {
-        parser.runIfNotNull(() -> {
-            tableEditor.addColumn(columnDefinitionListener.getColumn());
-        }, tableEditor, columnDefinitionListener);
-        super.exitColumnDeclaration(ctx);
-    }
-
-    @Override
-    public void enterPrimaryKeyTableConstraint(MySqlParser.PrimaryKeyTableConstraintContext ctx) {
-        parser.runIfNotNull(() -> {
-            parser.parsePrimaryIndexColumnNames(ctx.indexColumnNames(), tableEditor);
-        }, tableEditor);
-        super.enterPrimaryKeyTableConstraint(ctx);
-    }
-
-    @Override
-    public void enterUniqueKeyTableConstraint(MySqlParser.UniqueKeyTableConstraintContext ctx) {
-        parser.runIfNotNull(() -> {
-            if (!tableEditor.hasPrimaryKey() && parser.isTableUniqueIndexIncluded(ctx.indexColumnNames(), tableEditor)) {
-                parser.parsePrimaryIndexColumnNames(ctx.indexColumnNames(), tableEditor);
-            }
-        }, tableEditor);
-        super.enterUniqueKeyTableConstraint(ctx);
     }
 
     @Override
